fix(reports): AR aging (/ar-aging, /detail, /customer/{id}) rebuilt on compute_ar_outstanding -- same response contract
Measured 23 Sep: get_ar_aging_summary/detail/customer read sales_invoices.paid_amount (column does not
exist) -> all three routes 500 on every call. The chat's "umur piutang" action calls /api/reports/ar-aging
(owner rule: chat code untouched -- fix the ROUTE the chat already calls). Even when the column existed
it was the cache, not the ledger (Iron Law 1/16).

V298: ar_aging_open_invoices(tenant, as_of) over compute_ar_outstanding (invoice rows, outstanding > 0,
invoice_date <= as_of, buckets exactly as before); the three functions rebuilt on it with the SAME result
columns, money NUMERIC(18,2) (was ::BIGINT). get_ar_aging_customer now takes the tenant (was id-only).
Synthetic unapplied-CN rows (invoice_id NULL) are not bucketed; detail keeps per-customer rows only (as
before). Router: customer call passes tenant; a PAST as_of -> 422 "Umur piutang per tanggal lampau belum
tersedia; saldo dihitung dari jurnal per hari ini." (compute_ar_outstanding has no as-of date -- ageing
today's balances as if historical would be a confident wrong number). Models: 18 money fields int -> float;
KEYS unchanged.
DEPLOY: code first, then V298 (either order has no regression window: the old functions 500 anyway).

Gate (scratch; 3 invoices posted THROUGH the SI router with due dates in current / 1-30 / 120+, qty 1.5 ->
sen; B's cache made stale amount_paid=40,000): expectation = compute_ar_outstanding read directly and
bucketed in the gate; contract = field names parsed (ast) from the BASE tree's models.
  base RED  : 500 x3
  sab RED   : V298 with the outstanding source swapped to the CACHE -> B missing, 1-30 bucket 0 vs 55,552.50,
              customer total 74,070 vs 111,105
  new GREEN : keys == base contract (summary, detail item, customer + item); buckets, detail total and each
              invoice's balance == ledger, sen kept (18,517.50); past as_of -> 422

+++ backend/api_gateway/app/schemas/aging_reports.py
@@ -57,13 +57,13 @@ class AgingBracketsResponse(AgingBracketsBase):
 # ============================================
 
 class ARAgingSummary(BaseModel):
-    total_current: int
-    total_1_30: int
-    total_31_60: int
-    total_61_90: int
-    total_91_120: int
-    total_over_120: int
-    grand_total: int
+    total_current: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_1_30: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_31_60: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_61_90: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_91_120: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_over_120: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    grand_total: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
     overdue_count: int
 
 
@@ -77,13 +77,13 @@ class ARAgingDetailItem(BaseModel):
     customer_id: UUID
     customer_name: str
     customer_code: Optional[str] = None
-    current_amount: int
-    days_1_30: int
-    days_31_60: int
-    days_61_90: int
-    days_91_120: int
-    days_over_120: int
-    total_balance: int
+    current_amount: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    days_1_30: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    days_31_60: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    days_61_90: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    days_91_120: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    days_over_120: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    total_balance: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
     oldest_invoice_date: Optional[date] = None
     invoice_count: int
 
@@ -99,9 +99,9 @@ class ARCustomerAgingItem(BaseModel):
     invoice_number: str
     invoice_date: date
     due_date: date
-    total_amount: int
-    paid_amount: int
-    balance: int
+    total_amount: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    paid_amount: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
+    balance: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
     days_overdue: int
     aging_bucket: str
 
@@ -111,7 +111,7 @@ class ARCustomerAgingResponse(BaseModel):
     customer_name: str
     as_of_date: date
     items: List[ARCustomerAgingItem]
-    total_balance: int
+    total_balance: float  # V298: NUMERIC(18,2) dari jurnal; int membuang sen
 
 
 # ============================================
